+++ include/moti/memory/inl/stack_allocator.inl
@@ -3,18 +3,18 @@
 namespace moti {
     namespace memory {
 
-        template <size_t MaxSize, size_t Align>
+		template <uint32_t MaxSize, uint32_t Align>
         StackAllocator<MaxSize, Align>::StackAllocator()
             : m_pointer(m_data) {
 
         }
 
-        template <size_t MaxSize, size_t Align>
-        Block StackAllocator<MaxSize, Align>::allocate(size_t _bytes) {
+		template <uint32_t MaxSize, uint32_t Align>
+        Block StackAllocator<MaxSize, Align>::allocate(uint32_t _bytes) {
             Block result;
             if (_bytes == 0) return result;
 
-            const size_t alignedLength = roundToAlignment(Align, _bytes);
+            const uint32_t alignedLength = roundToAlignment(Align, _bytes);
             // out of mem
             if (m_pointer + alignedLength > m_data + MaxSize) {
                 return result;
@@ -27,7 +27,7 @@ namespace moti {
         }
 
 
-        template <size_t MaxSize, size_t Align>
+        template <uint32_t MaxSize, uint32_t Align>
         void StackAllocator<MaxSize, Align>::deallocate(Block& _block) {
             // nullptr
             if (!_block) return;
@@ -44,8 +44,8 @@ namespace moti {
             _block.reset();
         }
 
-        template <size_t MaxSize, size_t Align>
-        bool StackAllocator<MaxSize, Align>::reallocate(Block& _block, size_t _bytes) {
+        template <uint32_t MaxSize, uint32_t Align>
+        bool StackAllocator<MaxSize, Align>::reallocate(Block& _block, uint32_t _bytes) {
             if (_block.m_length == _bytes) return true;
 
             if (_bytes == 0u) return true;
@@ -55,7 +55,7 @@ namespace moti {
                 return true;
             }
 
-            const size_t alignedLength = roundToAlignment(Align, _bytes);
+            const uint32_t alignedLength = roundToAlignment(Align, _bytes);
 
             if (_block.m_length > _bytes) {
                 _block.m_length = alignedLength;
@@ -75,13 +75,13 @@ namespace moti {
             return false;
         }
 
-        template <size_t MaxSize, size_t Align>
+        template <uint32_t MaxSize, uint32_t Align>
         bool StackAllocator<MaxSize, Align>::owns(const Block& _block) const {
             return _block && (_block.m_ptr >= m_data && _block.m_ptr < m_data + MaxSize);
         }
 
 
-        template <size_t MaxSize, size_t Align>
+        template <uint32_t MaxSize, uint32_t Align>
         bool StackAllocator<MaxSize, Align>::isLastUsedBlock(const Block& _b) {
             return (static_cast<char *>(_b.m_ptr) + _b.m_length == m_pointer);
         }
